Split DoubleSeq.hs
