+++ .gitignore
@@ -2,4 +2,4 @@ dist/
 dist-newstyle/
 .cabal-sandbox/
 cabal.sandbox.config
-
+cabal.project.local
